v1.36
改进协程栈入口函数，使栈结构定义不依赖此函数的编译实现细节。
linux 64位下测试通过。

+++ Dzcot/DzList.h
@@ -15,7 +15,7 @@ extern "C"{
 #endif
 
 #define MEMBER_BASE( memAddr, type, mem )\
-    ( (type*)( (char*)(memAddr) - (int)(&((type*)0)->mem ) ) )
+    ( (type*)( (char*)(memAddr) - (size_t)(&((type*)0)->mem ) ) )
 
 inline void InitDList( DzDList* list )
 {
